+++ components/operations/meta-tokens/create.tsx
@@ -26,7 +26,7 @@ import { useAccount } from '@useelven/core';
 import { OperationsTokenIdInput } from '@/components/operations/operations-tokenid-input';
 
 const formSchema = z.object({
-  collectionTokenId: z.string().min(1, 'The field is required'),
+  tokenId: z.string().min(1, 'The field is required'),
   name: z.string().min(1, 'The field is required'),
   initialQuantity: z
     .string()
@@ -47,7 +47,7 @@ export const Create = ({ triggerTx, close }: OperationContentProps) => {
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: {
-      collectionTokenId: '',
+      tokenId: '',
       name: '',
       initialQuantity: '',
       attributes: '',
@@ -55,13 +55,13 @@ export const Create = ({ triggerTx, close }: OperationContentProps) => {
   });
 
   const onSubmit = ({
-    collectionTokenId,
+    tokenId,
     name,
     initialQuantity,
     attributes,
   }: z.infer<typeof formSchema>) => {
     const args: TypedValue[] = [
-      BytesValue.fromUTF8(collectionTokenId.trim()),
+      BytesValue.fromUTF8(tokenId.trim()),
       new BigUIntValue(new Bignumber(initialQuantity)),
       BytesValue.fromUTF8(name.trim()),
       BytesValue.fromUTF8(''),