+++ components/operations/non-fungible-tokens/change-attributes.tsx
@@ -60,29 +60,29 @@ export const ChangeAttributes = ({
   }: z.infer<typeof formSchema>) => {
     try {
       // TODO: replace with useElven useApiCall when ready to handle such cases
-      const tokenOnNetwork = await axios.get<{ nonce: number; ticker: string }>(
-        `${apiAddress}/nfts/${tokenId.trim()}`,
-        {
-          headers: {
-            'Content-Type': 'application/json',
-            Accept: 'application/json',
-          },
-        }
-      );
+      const tokenOnNetwork = await axios.get<{
+        nonce: number;
+        collection: string;
+      }>(`${apiAddress}/nfts/${tokenId.trim()}`, {
+        headers: {
+          'Content-Type': 'application/json',
+          Accept: 'application/json',
+        },
+      });
 
       const nonce = tokenOnNetwork?.data?.nonce;
-      const collectionTicker = tokenOnNetwork?.data?.ticker;
+      const collectionId = tokenOnNetwork?.data?.collection;
 
       // TODO: show the error in the transaction status modal
-      if (!nonce || !collectionTicker) {
+      if (!nonce || !collectionId) {
         console.error(
-          "Can't read the nonce or/and collection ticker of the token, using MultiversX API!"
+          "Can't read the nonce or/and collection id of the token, using MultiversX API!"
         );
         return;
       }
 
       const args: TypedValue[] = [
-        BytesValue.fromUTF8(collectionTicker.trim()),
+        BytesValue.fromUTF8(collectionId.trim()),
         new BigUIntValue(new Bignumber(nonce)),
         BytesValue.fromUTF8(attributes.trim()),
       ];
@@ -107,7 +107,7 @@ export const ChangeAttributes = ({
       close();
     } catch (e) {
       console.error(
-        "Can't read the nonce or/and collection ticker of the token, using MultiversX API!",
+        "Can't read the nonce or/and collection id of the token, using MultiversX API!",
         e
       );
     }