+++ components/operations/non-fungible-tokens/create.tsx
@@ -26,7 +26,7 @@ import { useAccount } from '@useelven/core';
 import { OperationsTokenIdInput } from '@/components/operations/operations-tokenid-input';
 
 const formSchema = z.object({
-  collectionTokenId: z.string().min(1, 'The field is required'),
+  tokenId: z.string().min(1, 'The field is required'),
   name: z.string().min(1, 'The field is required'),
   royalties: z
     .string()
@@ -54,7 +54,7 @@ export const Create = ({ triggerTx, close }: OperationContentProps) => {
   const form = useForm<z.infer<typeof formSchema>>({
     resolver: zodResolver(formSchema),
     defaultValues: {
-      collectionTokenId: '',
+      tokenId: '',
       name: '',
       royalties: '',
       uris: '',
@@ -64,7 +64,7 @@ export const Create = ({ triggerTx, close }: OperationContentProps) => {
   });
 
   const onSubmit = ({
-    collectionTokenId,
+    tokenId,
     name,
     royalties,
     uris,
@@ -72,7 +72,7 @@ export const Create = ({ triggerTx, close }: OperationContentProps) => {
     hash,
   }: z.infer<typeof formSchema>) => {
     const args: TypedValue[] = [
-      BytesValue.fromUTF8(collectionTokenId.trim()),
+      BytesValue.fromUTF8(tokenId.trim()),
       new BigUIntValue(new Bignumber(1)),
       BytesValue.fromUTF8(name.trim()),
       new BigUIntValue(new Bignumber(Number(royalties) * 100 || 0)),