+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+### [0.12.1](https://github.com/xdevguild/buildo.dev/releases/tag/v0.12.1) (2024-01-02)
+- important bugfix - the ticker is not always the same (API) as token/collection id
+- important bugfix - fix property name
+
 ### [0.12.0](https://github.com/xdevguild/buildo.dev/releases/tag/v0.12.0) (2024-01-01)
 - improve roles and properties selectors to make them less confusing
 - use tokenId selectors for fungible ids and collections